scripts/requirements.txt
aiohttp
python-dotenv
typing-extensions
pyyaml
pytest
pytest-asyncio
pytest-cov
mypy
aiofiles
types-aiofiles
attrs